fix sleep_now in aviator

+++ apps/utils/datetime.py
@@ -2,5 +2,5 @@
 import time
 
 
-def sleep_now(seconds: int):
+def sleep_now(seconds: float):
     time.sleep(seconds)
